Add Niche implementation for a test struct
This struct has two fields that have niches available, and the ‘Niche’
implementation explicitly chooses to use the second one.  I've been
careful to write the implementation so that it doesn't need to refer to
the field's type by name.  That will make it easier to write a proc
macro to implement this automatically for a struct.