Support generic structs in derive(Niche)
Our derive macro now works with generic struct types.  We copy over any
type parameters and where clauses in the struct definition.  We add an
additional where clause that ensures that the type of the struct's niche
field implements `Niche`.  (That might involve some of the type
parameters; it might not!)